--- sherlock/internal/slack/escape.go
@@ -1,15 +1,0 @@
-package slack
-
-import "strings"
-
-func EscapeText(text string) string {
-	replacements := map[string]string{
-		"&": "&amp;",
-		"<": "&lt;",
-		">": "&gt;",
-	}
-	for k, v := range replacements {
-		text = strings.ReplaceAll(text, k, v)
-	}
-	return text
-}
